% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airports.R
\docType{data}
\name{airports}
\alias{airports}
\title{US Airports}
\format{A data frame with 2671 rows and 4 variables:
\describe{
  \item{\code{AIRPORT}}{integer, the 3-letter code of an airport}
  \item{\code{AP_FULL_NAME}}{integer, the full name of an airport}
  \item{\code{LATITUDE}}{double, geographic latitude}
  \item{\code{LONGITUDE}}{double, geographic longitude}
}}
\usage{
airports
}
\description{
Information on all airports in the United States
}
\details{
Geographic nodes data frame for the Southwest network data.  For more information on edges of
network, see ?Southwest
}
\keyword{datasets}
