% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoNet.R
\name{geoNet}
\alias{geoNet}
\title{Creates a graph of a geographic network}
\usage{
geoNet(df, geodata = NULL, node_color = "red", node_size = 3,
  edge_color = "navy", edge_size = 4, name = "name", lat = "lat",
  lng = "lng", orig = "origin", dest = "destination")
}
\arguments{
\item{df}{This is the primary data set. It corresponds to the edges of the network.
It accepts an igraph or tidygraph object, or data frames or an adjacency matrices.
If this is an igraph or tidygraph object, the geodata parameter would be ignored
as all the node attributes (including the geographic data) would be extracted from
the object itself. It this is a data frame or and adjacency matrix, then geodata
cannot be NULL.}

\item{geodata}{This is a data frame that describes the attributes of the nodes,
including its geographic attributes. Default: NULL}

\item{node_color}{This accepts a character vector that is either a color or the
name of the variable amongst the node attributes that should be used to decide the
color of each node. The variable name should correspond to a variable that is either
a factor variable or is numeric. Default: 'red'}

\item{node_size}{This accepts either an integer or a character vector that is
the name of the variable amongst the node attributes that should be used to decide
the size of each node. The variable name should correspond to a variable that is
numeric in nature. Default: 3}

\item{edge_color}{This accepts a character vector that is either a color or the
name of the variable amongst the edge attributes that should be used to decide the
color of each edge The variable name should correspond to a variable that is either
a factor variable or is numeric. Default: 'navy'}

\item{edge_size}{This accepts either an integer or a character vector that is
the name of the variable amongst the edge attributes that should be used to decide
the width of each edge. The variable name should correspond to a variable that is
numeric in nature. Default: 4}

\item{name}{This accepts a character vector that is the name of the variable
amongst the node attributes that describes the name of each node. Default: 'name'}

\item{lat}{This accepts a character vector that is the name of the variable
amongst the node attributes that describes the latitude of each node. Default: 'lat'}

\item{lng}{This accepts a character vector that is the name of the variable
amongst the node attributes that describes the longitude of each node. Default: 'lng'}

\item{orig}{This accepts a character vector that is the name of the variable
amongst the edge attributes that describes one end of each edge. Default: 'origin'}

\item{dest}{This accepts a character vector that is the name of the variable
amongst the edge attributes that describes the second end of each edge. Default: 'destination'}
}
\value{
This function returns an interactive map with the network on it. This map is of the
class Leaflet.
}
\description{
\code{geoNet} accepts a network in the form of a set of nodes and
edges, where each node corresponds to an actual geographic location. It returns an
interactive geographic map with this network overlaid on it. It also
allows the user to specify other variables according that modify the color and size
of the nodes and edges of the network.
}
\details{
If users wish to customize their geoNet graph based on measures of network centrality, unless they
have this information on centrality measures for their network that are already featured in the
nodes/edges tables of the network, users should input their network into the baseNet function and then
use the tidygraph "data" output as the data argument for the geoNet function, making sure to specify
the modified node/edge arguments as necessary.
}
\examples{
\dontrun{
data(airports)
data(Southwest)

sw_nodes = unique(c(Southwest$ORIGIN, Southwest$DEST))
k2 = airports$AIRPORT \%in\% sw_nodes
airports=as.data.frame(airports[k2,])
flights<-subset(Southwest, select=c(3,4,2))

m<-geoNet(df = flights, geodata = airports, node_color = "red", node_size = 3,
edge_color = 'cornflowerblue', edge_size = 4, name = "AIRPORT", lat = "LATITUDE", lng = "LONGITUDE",
orig = "ORIGIN", dest = "DEST")

m
}
}
