% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageNet-package.R
\docType{package}
\name{imageNet-package}
\alias{imageNet}
\alias{imageNet-package}
\title{imageNet: Produces a concise, customizable visualization of a provided network.}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/kthomasfranz/imageNet/issues}
}

}
\author{
\strong{Maintainer}: Robert Kabacoff \email{rkabacoff@wesleyan.edu}

Authors:
\itemize{
  \item Kaitlyn Thomas-Franz \email{kthomasfranz@wesleyan.edu}
  \item Dhhyey Shah \email{dshah@wesleyan.edu}
}

}
\keyword{internal}
