% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageNet.R
\name{imageNet}
\alias{imageNet}
\title{Creates an image of network in R}
\usage{
imageNet(data, nodes = NULL, layout = "lgl", label, directed = F,
  node_color, node_size, edge_color, edge_size)
}
\arguments{
\item{data}{A network object that can be converted to a tidygraph tbl_graph
(acceptable classes include 'data.frame', 'igraph', 'list', 'matrix', and 'network').
See ?tbl_graph for more information about accepted classes of networks.}

\item{nodes}{A data.frame containing information about the nodes in the graph, Default: NULL}

\item{layout}{The type of layout used to plot the network using ggraph. Either a valid string, a function, a matrix,
or a data.frame. See ?ggraph for more information. Default: 'lgl'}

\item{label}{The variable from the nodes table that should be used to label the nodes in
the plotted network.  Default: first column of the nodes table}

\item{directed}{Indicates if the constructed graph should be directed, Default: F}

\item{node_color}{PARAM_DESCRIPTION, Default: "cornflowerblue"}

\item{node_size}{PARAM_DESCRIPTION, Default: 8}

\item{edge_color}{PARAM_DESCRIPTION, Default: "slategrey"}

\item{edge_size}{PARAM_DESCRIPTION, Default: 0.6}
}
\value{
list containing data in the form of a tidygraph & ggmap of network
}
\description{
\code{imageNet} helps provide users with clear, concise visualizations of
their networks so that they can perform effective analysis.
}
\details{
The \code{imageNet} function converts the user's network into a \code{tidygraph} object
and plots the network using \code{ggmap}. In order to operate this fucntion effectively,
it is expected that the user has an elementary understanding of network analysis and knows how
to perform a basic evaluation of some type of class of networks
(data.frame, igraph, network, etc.) in R.
}
\examples{
library(igraphdata)
data("karate")
imageNet(karate, label=name, layout="nicely")
}
