% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Southwest.R
\docType{data}
\name{Southwest}
\alias{Southwest}
\title{SouthWest Airline Routes, 1987-97}
\format{A data frame with 236 rows and 4 variables:
\describe{
  \item{\code{CARRIER}}{integer, the airline carrier, which is always WN (SouthWest)}
  \item{\code{YEAR}}{integer, lists whether the link between nodes/airports existed only in 1987,
  only in 1997, or both periods (“ALWAYS”)}
  \item{\code{ORIGIN}}{integer, the 3-letter codes of the origin airport}
  \item{\code{DEST}}{integer, the 3-letter codes of the destination airport}
}}
\source{
US Bureau of Transportation Statistics: Airline On-Time Performance Database
}
\usage{
Southwest
}
\description{
Network of SouthWest airlines routes between 1987 and 1997
}
\details{
Edges data frame for Southwest network data.
}
\keyword{datasets}
